refactor: extension 분리 및 주석제거
- extension 분리 파일을 만들어 폴더로 그룹화
    - NumberFormatter 적용 부분을 extension 으로 분리
    - UIImageView 이미지url을 받아서 이미지를 load하는 로직 분리
    - UILabel에 취소선 추가하는 로직 분리
- 불필요한 주석 제거
